+++ app/src/main/java/com/example/myapplication/bluedot_4/view/EdgeBadgeImageView.java
@@ -68,12 +68,12 @@ public class EdgeBadgeImageView extends ImageView implements IEdgeBadgeView, IEd
     }
 
     @Override
-    public IEdgeBadgeView setBadgePadding(float padding, boolean isDpValue) {
+    public IEdgeBadgeView setBadgePadding(int padding, boolean isDpValue) {
         return mEdgeBadgeViewHelper.setBadgePadding(padding, isDpValue);
     }
 
     @Override
-    public IEdgeBadgeView setMargin(float horizontalMargin, float verticalMargin, boolean isDpValue) {
+    public IEdgeBadgeView setMargin(int horizontalMargin, int verticalMargin, boolean isDpValue) {
         return mEdgeBadgeViewHelper.setMargin(horizontalMargin, verticalMargin, isDpValue);
     }
